# node/node-api-14.06.2023/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "Study Rest Api and swagger"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieves a list of all users",
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/add": {
      "post": {
        "summary": "Create a new user",
        "description": "Creates a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            },
            "description": "User object"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "examples": {
          "example-1": {
            "summary": "Example request",
            "value": {
              "id": 5,
              "name": "your_username",
              "actualName": "your_actual_name",
              "displayName": "AA"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a specific user",
        "description": "Retrieves information about a specific user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "put": {
        "summary": "Update a specific user",
        "description": "Updates information of a specific user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          },
          {
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            },
            "description": "User object"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "examples": {
          "example-1": {
            "summary": "Example request",
            "value": {
              "id": "1",
              "name": "John Doe",
              "actualName": "John Doe",
              "displayName": "John Doe"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "actualName": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        }
      }
    }
  }
}